Made changes to compile and run on Linux with gcc.
1.  Created CMakeLists.txt to build with CMake on Linux.  It might also
work on other systems.

2.  stricmp and strnicmp were being used; they are missing from
gcc's standard library implementation.  (This is a well-known and
long-standing bug in gcc.)  So I added src/utils/StringStuff.{cpp,h}
with implementations of case-insensitive string compare functions
and changed all instances of stricmp and strnicmp to use them.

3.  Added StringStuff to the VC project files so hopefully it'll still
build on that tool (which I do not have).

4.  Fixed a variety of minor bugs where the code was not standard
C++, but presumably VC was permissive enough that it compiled anyway.
Hopefully VC will accept these changes; I don't have VC.  The most
common bug was assigning a wxString to a std::string in an old-style
initializer; switching to the new initializer style fixed it.

5.  Added a few missing '#include' statements.

6.  Made some no-effect changes to silence compiler warnings, such as
adding "default: break;" to switch statements where some of the values
were not handled.

7.  gcc is missing acosf, cosf, sinf, atan2f, and fabsf.  (This is
a well-known and long-standing bug in gcc.)  Luckily, these are
just alternative names for existing functions, so I switched to the
other names.

8.  There were many locations where BS&OS tried to open files
with filenames where the path separator character was backslash.
I changed them to use forward slash instead.  This shouldn't affect
Windows execution, because Windows accepts either type of slash as a
path separator.  I tried my best to not let the slash style propagate
into data files, but I may not have been successful.

9.  There were a few locations where filenames were being converted to
lowercase before the file was opened.  As Linux is case sensitive, this
didn't work.  So I changed all of these instances so that the filename
is not converted.

10.  On Linux, the wxWidgets headers "#include <X11/X.h>", which is an
ancient legacy header file containing a large number of poorly-named
"#define" statements.  I feel that this is a bug in wxWidgets.  There
was a name clash with a function name in the FBX SDK.  So I moved the
"#include <fbxsdk.h>" line from FBXWrangler.h into FBXWrangler.cpp
so that FBX and wxWidgets headers would never be included by the
same file.  To keep FBXWrangler working, I needed to move more of
the implementation details from the header into the source file.
This necessitated employing the private-implementation idiom.

11.  The C integer-absolute-value function "abs" was being used in
TriFile.cpp on floating point values.  gcc produced a warning, and
I judged this highly likely to be a bug, so I fixed it by replacing
"abs" with "std::abs".

12.  Linux doesn't have a registry, so I used "#ifdef _WINDOWS"
around the handful of blocks of registry code.

13.  In BodySlideApp.h, there was some peculiar code for doing something
with number-key presses and Windows events.  I used "#ifdef _WINDOWS"
to disable this code for linux.

14.  A few files were missing trailing newline characters.
Linux doesn't like that, so I added them.

15.  Switched from old-style cast to "reinterpret_cast" a few places
in NormalsGenDialog to silence compiler complaints.

16.  "_snprintf_s" was used in a few places.  It's a Windows-only
function.  I replaced it with snprintf, which does the same thing
only better.

17.  The extension-loading code in GLExtensions.{h,cpp} used
wglGetProcAddress, which is Windows-specific.  I couldn't figure out
how to convert the code to Linux, so, for Linux, GL extensions are
now loaded by GLEW.  This new code would probably work fine on other
systems too, but it would introduce a dependence on the GLEW library.

18.  Three OpenGL API calls had capitalization errors.  I fixed them.

After all this, BodySlide and Outfit Studio build and run, apparently
correctly, on my Linux system.