Made all pathnames in files use backslashes.
1.  Added Readme-linux.txt, which describes the build process for Linux.

2.  Added ToOSSlashes and ToBackslashes functions to StringStuff.  They
convert slashes and backslashes to either the OS's path separator character
or backslashes, respectively.

3.  Added PathSepChar and PathSepStr constants to StringStuff.h.

4.  Made sure pathnames read in from files have backslashes converted
to the path separator character of the OS.  Made sure pathnames written
to files have forward slashes converted to backslashes.  This was done
for the following file types and data items:

material file: all texture pathnames
nif file:
	SetStringData (only for BODYTRI extra data on a shape)
	GetTextureSlot
	SetTextureSlot
slider set osp file:
	SliderSetInfo::SliderSet::DataFolder
	SliderSetInfo::SliderSet::OutputPath
	SliderSetInfo::SliderSet::Shape::DataFolder
	SliderSetInfo::SliderSet::Slider::Data
ref templates xml file: sourcefile attribute of RefTemplates::Template

I did not do this for the configuration file Config.xml, which has
a bunch, because I figured that file is never exchanged or used by
any other software.

5.  Generally made sure that both forward and back slashes are accepted
in pathnames internally, such as when splitting the filename off of
a path.