Set slider background color in Outfit Studio for Linux.
wxWidgets for gtk2 and gtk3 has a well-known bug that makes it ignore some
"bg" properties in xrc files.  This results in some widgets having
system-defined background colors.  In Outfit Studio, there's just one
place where this is a problem: the background colors of sliders.  This
change sets the background color in the code.