Fixed three bugs with the move brush (#238)
1.  When the move brush overlapped the centerline and x-mirroring was
on, vertices on the other side of the centerline and inside the brush
radius were not mirror-moving.  I added code to TB_Move::brushAction
to check whether each vertex is closer to the origin or mirror-origin.

2.  Rarely, when using the move brush, pointStartState is set
but pointEndState is not for some vertices.  This resulted in
garbage being sent to the slider diff update when slider editing
was on.  I couldn't figure out how this was happening (it should
be impossible), but I prevented the transfer of garbage by adding a
check to ActiveShapesUpdated.

3.  There was a sign bug in GLSurface::CollidePlane.  This bug
caused the move brush to often move points less than it should,
and sometimes in the opposite direction.  I fixed the sign bug.
Of the many places from which CollidePlane is called, the sign bug
was already compensated for in all places except UpdateBrushStroke
in the TBT_MOVE code, which was why it was leading to problems only
for the move brush.  I removed the sign-error compensation from all
of the other places from which CollidePlane is called.