Added weight normalization to weight brushes (#239)
* Added weight-balancing to weight brushes

This is the first of a sequence of commits to add automatic weight
balancing to Outfit Studio.  This commit just affects the three weight
brushes.  With this change, the weights for the bones on a vertex will
be maintained to have more or less the same sum as they did before.

A lot of the changes in this commit are designed to support the next
commit, which will add balance-bone selection to the GUI.