Fixed a scale bug in MatTransform::ToMatrix
In the function MatTransform::ToMatrix, the scale was being incorrectly
applied in calculating the result.  This would give an incorrect matrix
if the scale was not 1.  This function is called from two places:

- from NifFile::GetAbsoluteNodeTransform, where it shouldn't matter
because that function uses NiNodes, which generally have scale of 1.

- from AnimInfo::CalcShapeSkinBounds, where it was causing bone bounds
to be calculated incorrectly if xform had a scale other than 1.