Cleaned up the coordinate-system transform code.
Fixed scaling bug in MatTransform::ToMatrix.

Moved the implementation of MatTransform::ToEulerDegrees into Matrix3.
Made a radians version, ToEulerAngles.  Used ToEulerAngles
instead of ToEulerDegrees in Automorph::MeshFromNifShape and
GLSurface::AddMeshFromNif so that those functions don't have to
convert from degrees to radians when passing the values to glm::yawPitchRoll.

Added Matrix3::operator*(const Vector3 &).

Fixed bug in Matrix3::operator*(const Matrix3 &) where the identity
matrix was used instead of this matrix.

Fixed bug in Matrix3::operator*=(const Matrix3 &) where the two matrices
were swapped in the calculation.

Added Vector3::operator[].

Added Vector3 operator*(float, const Vector3 &).

Added Matrix3 Determinant, Invert, and Inverse functions.

Added MatTransform functions: ApplyTransform, InverseTransform, and
ComposeTransforms.  Used these functions in many places to simplify code.

Added NiAVObject functions GetTransformToParent and
SetTransformToParent and made everything except IO use these functions
rather than accessing "transform" directly.  I did this purely for
clarity, not for data hiding.

Simplified NifFile::GetAbsoluteNodeTransform.  Added a new name for it:
GetNodeTransformToGlobal.

Split functionality of NifFile::GetShapeBoneTransform into
GetShapeTransformFromGlobalToSkin and GetShapeTransformFromSkinToBone.
The old function now calls one of the two.  Same for SetShapeBoneTransform.

Added new name for NifFile::GetNodeTransform: GetNodeTransformToParent.
Same for SetNodeTransform.

Renamed AnimWeight::xform to xformSkinToBone.

Renamed AnimInfo::GetShapeBoneXForm to GetXFormSkinToBone.  Same for
SetShapeBoneXForm.

Replaced AnimBone's rot, trans, scale, localRot, and localTrans with
two MatTransforms: xformToGlobal and xformToParent.

Added MatTransform xformGlobalToSkin to AnimSkin.  This is set when
the Nif is read, but it's never written back out to the Nif.  Added a
line to OutfitProject::ResetTransforms to clear this transform, though
I can't see why calling ResetTransforms would ever be a good idea, so
I'm not sure if xformGlobalToSkin needs to be reset there.

AnimInfo::AddShapeBone is called from three places.  In all three
places, xformSkinToBone was being incorrectly set afterwards.  So I
added code to AddShapeBone to set xformSkinToBone to a good default
using the skeleton, and I eliminated the incorrect code from two of
the three callers (AddBoneRef and AddCustomBoneRef).  For the third
(AddShapeBoneAndXForm), the original code was supposed to copy the
xform from the base shape; I fixed it.

Eliminated AnimInfo::GetBoneXForm, which is no longer used.

Renamed AnimSkeleton::GetBoneTransform to GetBoneTransformToGlobal.

Made AnimInfo::WriteToNif use AnimWeight::xformSkinToBone in all cases
instead of recalculating it for one case.  The code for recalculating
it was incorrect, anyway (if the global-to-skin transform was messy).

Fixed probably harmless bug in AnimInfo::WriteToNif that was making
NifFile::SetShapeBoneWeights called if not FO4, when the proper
test was not BSShape.  (That is, I changed "if (!isFO4)" to "if
(!isBSShape)".)

Fixed likely bug in AnimInfo::CalcShapeSkinBounds where radius was
not being scaled.

Eliminated AnimSkeleton::GetSkinTransform, which was no longer being called.

Commented out lines in AnimSkeleton::GetBoneTransformToGlobal and
AnimInfo::WriteToNif that were setting scale to 1 for the skeleton's
bone-to-global transforms.  Normally, scale is already 1 for every
bone in the skeleton.  If they're not 1 for the skeleton, it's probably
for a good reason, so we shouldn't override that.

Cleaned up Automorph::MeshFromNifShape and GLSurface::AddMeshFromNif,
particularly for non-skinned meshes.  The transform for non-skinned
meshes is probably still wrong, but it should be closer to being right.

testing:
Loading CBBE_Body.nif and writing it back out: transforms unchanged.

Loading TypeGNoir_1.nif and writing it back out: transforms unchanged,
unlike before.

Add a bone, give it some weight, and write: transforms look good.

Copy bone weights to new bone and write: transforms look good.

Add a bone to TypeGNoir_1.nif, give it some weight, and write: transforms
look good.

For TypeGNoir_1.nif, for shapes with scale!=1, bounding spheres don't
appear correctly in NifSkope.  I don't know if this is a bug in NifSkope
or bad data in TypeGNoir_1.nif.  When Outfit Studio writes out this file,
the bounding spheres are completely different, and NifSkope still doesn't
display them correctly.  I don't know if this is a bug in NifSkope or
Outfit Studio.

Weight painting doesn't work right for shapes with scale!=1 in Outfit
Studio.  My best guess is that this is because of brush size not scaling.
Other brushes have similar problems.