Tweak new global-to-skin calc by using skeleton instead of nif.
I changed AnimSkin's constructor to get the bone-to-global transform
from the skeleton instead of the bone node's bone-to-parent transform.
This should make it not depend on any peculiarities of the nif file's
bone nodes.

I tested by loading "FO4 CBBEBody.nif", adding a bone, painting to it,
exporting it, loading it in NifSkope, and checking that the calculated
skin-to-bone transform was still right.