Implemented posing.
-Added GUI elements that only appear when the bone tab is selected
that allow the user to turn posing on and off, select a bone to pose,
adjust the six DOFs of the pose, and reset one or all bone poses.
Posing is always off when the bone tab is not selected.  (Changes to
OutfitStudio.{xrc,h,cpp}.)

-Added functions RotVecToMat and RotMatToVec to
lib/NIF/utils/Object3d.{h,cpp}.  These change a rotation vector into
a matrix and vice versa.

-Fixed bugs in Vector2::operator!= and Vector3::operator!=.

-Added member variables poseRotVec, poseTranVec, and xformPoseToGlobal
to AnimBone for keeping track of the pose.  Added function
AnimBone::UpdatePoseTransform, which updates xformPoseToGlobal
when the pose changes.  Added code to AnimInfo::LoadFromNif and
AnimBone::LoadFromNif to initialize xformPoseToGlobal.

-Added code to OutfitProject::GetLiveVerts for applying the pose when
posing is on.  Added a bool "bPose" to OutfitProject that controls
whether GetLiveVerts applies the pose.

-Added code to OutfitStudioFrame::ActiveShapesUpdated to use
GetLiveVerts to update the mesh when updating weights with posing on.

-Made wxGLPanel::ApplyUndoState and wxGLPanel::EndBrushStroke not
call OutfitProject::ApplyBoneScale if the bone-scale slider position
is 0.  I did this because the code in ApplyBoneScale resets the mesh
vertices, preventing weighting changes from appearing.  The code in
ApplyBoneScale seems very strange to me, so I was afraid to touch it.

Things to be aware of:

-I put no effort into making the GUI pretty.  It needs editing
to match the style of the rest of Outfit Studio.  GUI styling is a
matter of personal taste, and Outfit Studio is not my project, so I
would prefer not to do this editing myself.

-Someone other than me should take a careful look at the new
code for updating the visuals in OutfitStudioFrame::ApplyPose and
OutfitStudioFrame::ActiveShapesUpdated to make sure it's right.

-I don't think I handled the interaction between posing and bone
scaling well.  I would rather get rid of bone scaling.