Cleaned up the coordinate-system transform code (#250)
* Fixed a scale bug in MatTransform::ToMatrix

In the function MatTransform::ToMatrix, the scale was being incorrectly
applied in calculating the result.  This would give an incorrect matrix
if the scale was not 1.  This function is called from two places:

- from NifFile::GetAbsoluteNodeTransform, where it shouldn't matter
because that function uses NiNodes, which generally have scale of 1.

- from AnimInfo::CalcShapeSkinBounds, where it was causing bone bounds
to be calculated incorrectly if xform had a scale other than 1.

* Cleaned up the coordinate-system transform code.

Fixed scaling bug in MatTransform::ToMatrix.

Moved the implementation of MatTransform::ToEulerDegrees into Matrix3.
Made a radians version, ToEulerAngles.  Used ToEulerAngles
instead of ToEulerDegrees in Automorph::MeshFromNifShape and
GLSurface::AddMeshFromNif so that those functions don't have to
convert from degrees to radians when passing the values to glm::yawPitchRoll.

Added Matrix3::operator*(const Vector3 &).

Fixed bug in Matrix3::operator*(const Matrix3 &) where the identity
matrix was used instead of this matrix.

Fixed bug in Matrix3::operator*=(const Matrix3 &) where the two matrices
were swapped in the calculation.

Added Vector3::operator[].

Added Vector3 operator*(float, const Vector3 &).

Added Matrix3 Determinant, Invert, and Inverse functions.

Added MatTransform functions: ApplyTransform, InverseTransform, and
ComposeTransforms.  Used these functions in many places to simplify code.

Added NiAVObject functions GetTransformToParent and
SetTransformToParent and made everything except IO use these functions
rather than accessing "transform" directly.  I did this purely for
clarity, not for data hiding.

Simplified NifFile::GetAbsoluteNodeTransform.  Added a new name for it:
GetNodeTransformToGlobal.

Split functionality of NifFile::GetShapeBoneTransform into
GetShapeTransformFromGlobalToSkin and GetShapeTransformFromSkinToBone.
The old function now calls one of the two.  Same for SetShapeBoneTransform.

Added new name for NifFile::GetNodeTransform: GetNodeTransformToParent.
Same for SetNodeTransform.

Renamed AnimWeight::xform to xformSkinToBone.

Renamed AnimInfo::GetShapeBoneXForm to GetXFormSkinToBone.  Same for
SetShapeBoneXForm.

Replaced AnimBone's rot, trans, scale, localRot, and localTrans with
two MatTransforms: xformToGlobal and xformToParent.

Added MatTransform xformGlobalToSkin to AnimSkin.  This is set when
the Nif is read, but it's never written back out to the Nif.  Added a
line to OutfitProject::ResetTransforms to clear this transform, though
I can't see why calling ResetTransforms would ever be a good idea, so
I'm not sure if xformGlobalToSkin needs to be reset there.

AnimInfo::AddShapeBone is called from three places.  In all three
places, xformSkinToBone was being incorrectly set afterwards.  So I
added code to AddShapeBone to set xformSkinToBone to a good default
using the skeleton, and I eliminated the incorrect code from two of
the three callers (AddBoneRef and AddCustomBoneRef).  For the third
(AddShapeBoneAndXForm), the original code was supposed to copy the
xform from the base shape; I fixed it.

Eliminated AnimInfo::GetBoneXForm, which is no longer used.

Renamed AnimSkeleton::GetBoneTransform to GetBoneTransformToGlobal.

Made AnimInfo::WriteToNif use AnimWeight::xformSkinToBone in all cases
instead of recalculating it for one case.  The code for recalculating
it was incorrect, anyway (if the global-to-skin transform was messy).

Fixed probably harmless bug in AnimInfo::WriteToNif that was making
NifFile::SetShapeBoneWeights called if not FO4, when the proper
test was not BSShape.  (That is, I changed "if (!isFO4)" to "if
(!isBSShape)".)

Fixed likely bug in AnimInfo::CalcShapeSkinBounds where radius was
not being scaled.

Eliminated AnimSkeleton::GetSkinTransform, which was no longer being called.

Commented out lines in AnimSkeleton::GetBoneTransformToGlobal and
AnimInfo::WriteToNif that were setting scale to 1 for the skeleton's
bone-to-global transforms.  Normally, scale is already 1 for every
bone in the skeleton.  If they're not 1 for the skeleton, it's probably
for a good reason, so we shouldn't override that.

Cleaned up Automorph::MeshFromNifShape and GLSurface::AddMeshFromNif,
particularly for non-skinned meshes.  The transform for non-skinned
meshes is probably still wrong, but it should be closer to being right.

* Renamed tta to ttg and replaced word "absolute" with "global"

* Fixed FO4 global-to-skin transform.

In previous patches, I mistakenly assumed that the global-to-skin transform
for nif files with BSSkinBoneData (which has no skinTransform) was the
identity.  It's not.  Changes:

- Changed a few minor things, mostly comments, in lib/NIF to reflect the
fact that this transform is not the identity.

- Moved the implementation of AnimWeight and AnimSkin constructors from
Anim.h to Anim.cpp.

- Added code to the AnimSkin constructor to calculate xformGlobalToSkin
if it's not present in the nif file.

* Tweak new global-to-skin calc by using skeleton instead of nif.

I changed AnimSkin's constructor to get the bone-to-global transform
from the skeleton instead of the bone node's bone-to-parent transform.
This should make it not depend on any peculiarities of the nif file's
bone nodes.

I tested by loading "FO4 CBBEBody.nif", adding a bone, painting to it,
exporting it, loading it in NifSkope, and checking that the calculated
skin-to-bone transform was still right.

* Eliminated OutfitProject::AddShapeBoneAndXForm.

I believe the code in AddShapeBoneAndXForm for calculating the new
bone's shape-to-bone transform was not producing a different transform
than the default code in AnimInfo::AddShapeBone, so I eliminated it.