Fix: preserve custom bone skeleton
This patch makes Outfit Studio preserve the skeleton tree structure
of custom bones and new bones when nif files are loaded and saved.
This makes posing work correctly for custom bones, as well as being
a good idea in general.  The details:

-Changed NifFile::AddNode to take a third parameter, the parent
node, and return the node added instead of its block ID.  The node
is added as a child of the root node if the parent node is nullptr,
just as before, and that's the default value.

-Split AnimSkeleton::AddBone into two functions, AddStandardBone and
AddCustomBone, for clarity.  The functionality is unchanged.

-Added function AnimSkeleton::LoadCustomBoneFromNif.  Called this
function from AnimInfo::LoadFromNif instead of adding the bone
directly.  The new function recursively adds ancestors, thus preserving
the skeleton tree.

-Added function AnimBone::AddToNif.  This function recursively adds
the bone's ancestors to the nif, thus preserving the skeletal structure
for new bones.  Changed AnimInfo::WriteToNif to call AddToNif.

-Eliminated AnimBone::boneID.  It was being set but never used.

-Added function AnimBone::SetTransformBoneToParent, which sets
xformToParent and also updates the other transforms that depend on it.
Called this function everywhere that xformToParent used to be set.

-Eliminated AnimSkeleton::DisableCustomTransforms.  It was never called,
and its functionality was dubious.

-Eliminated AnimSkeleton::allowCustomTransforms.  Its value was never
changed, and its functionality was dubious.

Testing:

-I loaded BloodSkirt03_1.nif and checked that posing works.  I also
exported it and made sure no transforms or skeletal structure were
changed.

-I added a standard bone to BloodSkirt03_1.nif, exported, and checked
to make sure that the skeletal structure was correctly added and the
transforms were correct.

-I couldn't test the ability to preserve custom skeletal structure
on write to nif, because Outfit Studio currently lacks the ability
to add new custom bone skeletal structure.  Perhaps I'll do that next.

Questions:

-Should the output nif's skeletal structure be rearranged to match the
reference skeleton when writing out a nif file?  Currently, Outfit
Studio is careful not to add bone nodes or change their structure
unless necessary (and it's only necessary when adding a new bone).

-Should AnimInfo::WriteToNif update NiNode::transformToParent for
all bone nodes?  Currently it only updates it for bone nodes whose
parent is the root node, setting it to AnimBone::xformToGlobal.
If it should update it for all bone nodes, what should it do if the
nif's skeletal structure doesn't match the reference skeleton (which
it often does not)?