Improved add-custom-bone dialog.
Added "parent bone" and "rotation" UI elements to the add-custom-bone
dialog (dlgCustomBone) and implemented them.

Most of the changes are in OutfitStudioFrame::OnAddCustomBone.

Added function AnimBone::SetParentBone and called it when appropriate.

Modified OutfitProject::AddCustomBoneRef to take parent bone and a
full transformation instead of just the new bone's origin.

Added code for keeping track of the bone right-clicked on in the bones
tab; I called contextBone.  It's used in OnAddBone to automatically
navigate to the context bone in the skeleton tree.  It's used in
OnAddCustomBone to initialize the parent-bone choice box.