Partition simplification 1: use trueTriangles whenever possible.
This is step one of my simplification of the partition code.
I modified NifFile.cpp and OutfitStudio.cpp to use the partition's
trueTriangles instead of sometimes using trueTriangles and sometimes
using (index-mapped) "triangles".

The details:

- Added flag bMappedIndices to NiSkinPartition.  It determines whether
"triangles" is treated as using partition vertex indices or shape
vertex indices.  bMappedIndices is set or updated in:
	NiSkinPartition::Get
	NifFile::OptimizeFor
	NifFile::CreateSkinning

- Renamed NifFile::TriangulatePartitions to
NiSkinPartition::ConvertStripsToTriangles and moved it to Skin.cpp.

- Extracted code for converting between trueTriangles, mapped
triangles, and vertexMap from OptimizeFor, UpdateSkinPartitions,
and SetShapePartitions, creating the following new functions (mostly
in NiSkinPartition):
	GenerateTrueTrianglesFromMappedTriangles
	GenerateMappedTrianglesFromTrueTrianglesAndVertexMap
	GenerateVertexMapFromTrueTriangles
	PrepareTrueTriangles
	SetTrueTriangles
	PrepareVertexMapsAndTriangles
	ApplyMapToTriangles
	CalcMaxTriangleIndex

- Used the new functions so that data is calculated only when needed.
Note that, because of the way it works, UpdateSkinPartitions tends to
recalculate everything.

- Rewrote algorithm of GenerateMappedTriangles... (formerly part of
OptimizeFor) to be more efficient.

- Made GetShapePartitions, SetShapePartitions, and UpdateSkinPartitions
use trueTriangles.

- Deleted code from OutfitStudioFrame::ShowPartition for handling
mapped-index triangles.  It now assumes all triangles are "true"
triangles.

This update does not change any externally visible functionality.
Everything ought to work the same as before, perhaps a little faster.
Nevertheless, I tested it against SK and SSE nif files to make sure
partitions could be edited, saved, and reloaded.