Partition simplification 2: eliminate verts from PartitionItemData.
This commit is fairly straightforward.  I eliminated the member
variable "verts" from PartitionItemData in OutfitStudio.h, along with
all code for maintaining it.  I also eliminated this parameter from
SetShapePartitions and GetShapePartitions in NifFile.

Mostly this was a straight code deletion.  The main exception was
in the color code in ShowPartition, which needed to be adjusted.

Strangely, some of my last edits from the previous commit showed up
here instead.  git is dumb.