Partition simplification 3: triParts in OutfitStudio.cpp.
I eliminated the "tris" member variable of PartitionItemData
in OutfitStudio.h, replacing it with two things: an "index"
member of PartitionItemData, and a vector-of-ints "triParts" in
OutfitStudioFrame.

triParts is the same size as the shape's triangle list.  Each int
gives the index of the partition the triangle belongs to, or -1 if
none.  It should be impossible for any triangle to not belong to a
partition, but the code doesn't misbehave if an index is -1.

-Added code to OnAddPartition to determine the index of the new partition.

-Added code to OnPartitionApply to convert triParts back into triangle
lists for each partition.

-Added code to CreatePartitionTree to generate triParts from the
partition triangle lists.

-Simplified ShowPartition's code for updating the partition.

-Changed ShowPartition's color code to match new data format.

-Fixed a bug in lib/NIF/utils/Object3d.h: The "operator==" function
for Triangles was duplicated (defined twice).