Partition simplification 4: move triParts calc into NiSkinPartition
-Extracted the code for going to and from triParts from CreatePartitionTree
and OnPartitionApply into two new NiSkinPartition functions:
	GenerateTriPartsFromTrueTriangles
	GenerateTrueTrianglesFromTriParts

-Changed the interfaces of GetShapePartitions and SetShapePartitions to
use triParts instead of partition triangle lists.

-Added code to GetShapePartitions and SetShapePartitions to ensure that
partitionInfo is at least as big as the partition list.

-Added code to SetShapePartitions to ensure that every triangle belongs
to a partition.  Unassigned triangles are added to a new partition.
(I'm not sure if unassigned or misassigned triangles are even possible,
but it's better to be robust.)

-Added triParts to NiSkinInstance.  It isn't used by anything but
the two new functions yet.  It will be more thoroughly connected in
the next commit.

-Fixed bug in OnAddPartition that should have been part of the
previous commit.  (I forgot to save before compiling and committing.)