Refactoring of partition code (#260)
* Partition simplification 1: use trueTriangles whenever possible.

This is step one of my simplification of the partition code.
I modified NifFile.cpp and OutfitStudio.cpp to use the partition's
trueTriangles instead of sometimes using trueTriangles and sometimes
using (index-mapped) "triangles".

The details:

- Added flag bMappedIndices to NiSkinPartition.  It determines whether
"triangles" is treated as using partition vertex indices or shape
vertex indices.  bMappedIndices is set or updated in:
	NiSkinPartition::Get
	NifFile::OptimizeFor
	NifFile::CreateSkinning

- Renamed NifFile::TriangulatePartitions to
NiSkinPartition::ConvertStripsToTriangles and moved it to Skin.cpp.

- Extracted code for converting between trueTriangles, mapped
triangles, and vertexMap from OptimizeFor, UpdateSkinPartitions,
and SetShapePartitions, creating the following new functions (mostly
in NiSkinPartition):
	GenerateTrueTrianglesFromMappedTriangles
	GenerateMappedTrianglesFromTrueTrianglesAndVertexMap
	GenerateVertexMapFromTrueTriangles
	PrepareTrueTriangles
	SetTrueTriangles
	PrepareVertexMapsAndTriangles
	ApplyMapToTriangles
	CalcMaxTriangleIndex

- Used the new functions so that data is calculated only when needed.
Note that, because of the way it works, UpdateSkinPartitions tends to
recalculate everything.

- Rewrote algorithm of GenerateMappedTriangles... (formerly part of
OptimizeFor) to be more efficient.

- Made GetShapePartitions, SetShapePartitions, and UpdateSkinPartitions
use trueTriangles.

- Deleted code from OutfitStudioFrame::ShowPartition for handling
mapped-index triangles.  It now assumes all triangles are "true"
triangles.

This update does not change any externally visible functionality.
Everything ought to work the same as before, perhaps a little faster.
Nevertheless, I tested it against SK and SSE nif files to make sure
partitions could be edited, saved, and reloaded.

* Partition simplification 2: eliminate verts from PartitionItemData.

This commit is fairly straightforward.  I eliminated the member
variable "verts" from PartitionItemData in OutfitStudio.h, along with
all code for maintaining it.  I also eliminated this parameter from
SetShapePartitions and GetShapePartitions in NifFile.

Mostly this was a straight code deletion.  The main exception was
in the color code in ShowPartition, which needed to be adjusted.

Strangely, some of my last edits from the previous commit showed up
here instead.  git is dumb.

* Partition simplification 3: triParts in OutfitStudio.cpp.

I eliminated the "tris" member variable of PartitionItemData
in OutfitStudio.h, replacing it with two things: an "index"
member of PartitionItemData, and a vector-of-ints "triParts" in
OutfitStudioFrame.

triParts is the same size as the shape's triangle list.  Each int
gives the index of the partition the triangle belongs to, or -1 if
none.  It should be impossible for any triangle to not belong to a
partition, but the code doesn't misbehave if an index is -1.

-Added code to OnAddPartition to determine the index of the new partition.

-Added code to OnPartitionApply to convert triParts back into triangle
lists for each partition.

-Added code to CreatePartitionTree to generate triParts from the
partition triangle lists.

-Simplified ShowPartition's code for updating the partition.

-Changed ShowPartition's color code to match new data format.

-Fixed a bug in lib/NIF/utils/Object3d.h: The "operator==" function
for Triangles was duplicated (defined twice).

* Partition simplification 4: move triParts calc into NiSkinPartition

-Extracted the code for going to and from triParts from CreatePartitionTree
and OnPartitionApply into two new NiSkinPartition functions:
	GenerateTriPartsFromTrueTriangles
	GenerateTrueTrianglesFromTriParts

-Changed the interfaces of GetShapePartitions and SetShapePartitions to
use triParts instead of partition triangle lists.

-Added code to GetShapePartitions and SetShapePartitions to ensure that
partitionInfo is at least as big as the partition list.

-Added code to SetShapePartitions to ensure that every triangle belongs
to a partition.  Unassigned triangles are added to a new partition.
(I'm not sure if unassigned or misassigned triangles are even possible,
but it's better to be robust.)

-Added triParts to NiSkinInstance.  It isn't used by anything but
the two new functions yet.  It will be more thoroughly connected in
the next commit.

* Partition simplification 5: use triParts in lib/NIF

For this, last partition simplification commit, I searched for places
in lib/NIF where using triParts instead of trueTriangles could mean
simpler code or greater efficiency.  The significant changes:

-Added a function NiSkinPartition::PrepareTriParts that generates
triParts only when necessary.

-Changed calls of GenerateTriPartsFromTrueTriangles to calls of
PrepareTriParts.

-Added code in several places to clear triParts whenever trueTriangles
is changed such that triParts is no longer valid.

-Added code to NifFile::SetDefaultPartition to initialize bMappedIndices
and not set triangles incorrectly (bug fix).

-Added code to NifFile::PrepareData to trivially initialize triParts
for BSTriShapes.

-Deleted a lot of code from UpdateSkinPartitions that was intended
to do a flood-fill from each triangle in a partition, absorbing more
triangles into the partition.  Because the flood-fill queue was not
primed properly, this code was unused.  (bug fix)

-Fixed a bug in UpdateSkinPartitions where partBones was not being
re-indexed when a new partition was inserted due to a partition having
too many bones.

-Changed UpdateSkinPartitions to use triParts directly instead of
trueTriangles.  The git-diff output makes it look like I changed a
lot more than I actually did, because I reduced the indentation of the
main all-triangles loop (so it's no longer inside the partition loop).
I may have also increased the overall algorithm efficiency from O(n^2)
to O(n) by eliminating the search for each partition triangle in the
main triangle list.

-Added operator[] to Triangle.

Co-authored-by: ousnius <ousnius@users.noreply.github.com>