Fixed ImportFBX not setting up bones properly (#270)
Shapes from imported FBX files were not getting their skin-to-bone
transforms set.  Partly, this was because OutfitProject::ImportFBX
duplicated the code of AnimInfo::AddShapeBone instead of calling
AddShapeBone; as the code for AddShapeBone was updated, ImportFBX got
out of date.  So I deleted a lot of code and just called AddShapeBone.
But it's a mystery to me how the code in ImportFBX could have ever
worked correctly: it was not setting the skin-to-bone transforms at
all.  Perhaps it was relying on some old code in AnimInfo::WriteToNif
or something like that.

In addition, I spotted some bugs, and I fixed one:

NifFile::SetShapeBoneIDList was being called with bad data
in ImportFBX.  SetShapeBoneIDList wants a list of Nif block
IDs of the bones, but it was being passed the bone indices from
AnimSkin::boneWeights, which are never the same thing.  Perhaps this
was desirable for some reason; in any case, the correct data would
eventually be put in the NifFile by AnimInfo::WriteToNif.  I deleted
this bad call of SetShapeBoneIDList.