Segment simplification 1: triParts in OutfitStudioFrame.
I replaced std::set<uint> tris in (Sub)SegmentItemData in OutfitStudio.h
with an int partID and used triParts to keep track of triangle segment
membership.

This involved changes to most of the segment functions of OutfitStudioFrame.

I also changed how segments are colored.  They are colored the same
as partitions, now.