Added copy-bone-weight option to also copy skin transform.
-Added a checkbox with explanatory text to the copy-bone-weights dialog
(dlgCopyWeights) for copying the reference's global-to-skin transform,
and implemented it.

-Added a function OutfitStudioFrame::CalcCopySkinTransOption that
calculates whether it makes sense to show the option and what the default
value for the option should be.

-Added function AnimInfo::ChangeGlobalToSkinTransform that sets the
transform and updates all the skin-to-bone transforms.

-Added IsNearlyEqualTo functions to MatTransform, Matrix3, and
Vector3, and added a FloatsAreNearlyEqual function.  These are in
lib/NIF/utils/Object3d.h.

Testing: I loaded FO4 CBBEBody, exported it to obj, imported it from
obj, merged in the CBBEBody as a reference, and copied bone weights.
The option appeared and was checked.  I then saved and checked that the
two shapes had identical skin-to-bone transforms for Spine2 in NifSkope.

When the copy-transforms option is hidden in the dialog, it leaves
an empty space.  Perhaps there's some good way to make the dialog lay
itself out again, but I'm no expert in wxWidgets, so I left that bug
to someone else to fix.