Added option to transform to global on obj output
I added code that checks for a non-identity global-to-skin transform
for a shape when exporting to obj.  If it is, the user is asked if
they wish to transform the geometry to global coordinates.  The
transform is applied in OutfitProject::ExportOBJ.

Testing: In OS, I exported FO4 CBBEBody to obj.  The option appeared; I
clicked "Yes".  Then I cleared the project and imported it.  It showed
up in the wrong place, higher.  (Is this because of some transform
that's applied only when in FO4 mode?  I imported it in SSE mode, and
it displayed fine.)  I merged in FO4 CBBEBody; they did not visibly
match locations.  (Perhaps this is because the global-to-skin transform
isn't obtained correctly in GLSurface::AddMeshFromNif?  Maybe a bug
in that function.)  I tried "copy bone weights".  The copy-transform
checkbox appeared, but the default was false (both as expected).
I tried the copy-transform option both ways, and neither way did it
copy any bone weights.  (Perhaps this is a bug in CopyBoneWeights
that makes it not take the global-to-skin transform into account.
Does it only work for shapes with the same skin CS?)  So I painted some
weight onto the obj-imported shape and exported both to nif.  I loaded
it in NifSkope, and the shapes lined up perfectly, with do-skinning on.

My conclusion is that the transform-to-global code works correctly,
despite all the problems in Outfit Studio.

It looks like Outfit Studio does some peculiar things when in FO4 mode.
Am I right, that these are bugs in OS?  Or am I totally confused?