Implemented undo for delete vertices and separate vertices.
The "Delete Vertices..." and "Separate Vertices..." commands are now
undoable, except for those portions of the commands that delete or
duplicate shapes.

Changes in lib/NIF:

- Added functions:
	BSTriShape::GetEyeData
	BSTriShape::SetTangentData
	BSTriShape::SetBitangentData
	BSTriShape::SetEyeData
	NifFile::GetTangentsForShape
	NifFile::GetBitangentsForShape
	NifFile::GetEyeDataForShape
	NifFile::SetTangentsForShape
	NifFile::SetBitangentsForShape
	NifFile::SetEyeDataForShape
	(NifUtil) InsertVectorIndices
	(NifUtil) GenerateIndexExpandMap

- Made the various Create functions use "const" for their input parameters.

- Fixed Create functions so that they won't crash if passed a nullptr for
tris.

- Fixed NifFile::GetNormalsForShape not checking if shape has normals.

- Made NifFile::SetVertsForShape call Create, instead of aborting,
if the number of vertices has changed.

Changes in src/components:

- Added functions:
	AnimSkin::InsertVertexIndices
	Automorph::InsertVertexIndices
	Automorph::GetVertexDiffs
	Automorph::SetVertexDiffs
	DiffDataSets::InsertVertexIndices
	DiffDataSets::GetVertexDiffs
	DiffDataSets::SetVertexDiffs

- Simplified functions using lib/NIF/NifUtil.h functions:
	AnimInfo::DeleteVertsForShape
	DiffDataSets::DeleteVerts

- Added new undo type UT_MESH with supporting data structures in
UndoState.h.

Changes in OutfitProject and OutfitStudio:

- Added functions:
	OutfitProject::CollectVertexData
	OutfitProject::CollectTriangleData
	OutfitProject::PrepareDeleteVerts
	OutfitProject::ApplyShapeMeshUndo

- Changed OnDeleteVerts, OnSeparateVerts, and ApplyUndoState to
handle the new undo data and call the new OutfitProject functions.

The core functionality of this update is in OutfitProject.  Most of
the rest is just to support those new functions.  The first three
functions fill in UndoStateShape, and the fourth uses that data.
ApplyShapeMeshUndo is huge, as it grabs all of the data, edits it,
and puts it back into the various nif, anim, and diff data structures.

This was "plan B" of the two plans I presented in issue thread 254.
I got the feeling, while implementing this, that plan A would have
been a little simpler in addition to being faster.