Copy skin transforms with weights and added coordinates tab in shape props (#275)
* Added copy-bone-weight option to also copy skin transform.

-Added a checkbox with explanatory text to the copy-bone-weights dialog
(dlgCopyWeights) for copying the reference's global-to-skin transform,
and implemented it.

-Added a function OutfitStudioFrame::CalcCopySkinTransOption that
calculates whether it makes sense to show the option and what the default
value for the option should be.

-Added function AnimInfo::ChangeGlobalToSkinTransform that sets the
transform and updates all the skin-to-bone transforms.

-Added IsNearlyEqualTo functions to MatTransform, Matrix3, and
Vector3, and added a FloatsAreNearlyEqual function.  These are in
lib/NIF/utils/Object3d.h.

* Added option to transform to global on obj output

I added code that checks for a non-identity global-to-skin transform
for a shape when exporting to obj.  If it is, the user is asked if
they wish to transform the geometry to global coordinates.  The
transform is applied in OutfitProject::ExportOBJ.

* Added global-to-skin-transform editor to shape properties.

In addition to the new editor, I also:

-Changed the wording of the output-to-obj question about transforming
to global to make it clear that it's not recommended.

-Added a checkbox for transforming the geometry to the copy-weight
dialog box.

-Added function NifFile::CalcShapeTransformGlobalToSkin that calculates
the transform for FO4 shapes as well as older shapes.  This function is
used in BodySlide's preview window.

-Separated matModel global-to-skin transform code out of
GLSurface::AddMeshFromNif and put it in GLSurface::SetSkinModelMat
so it can be called with the right transform from OutfitStudio and
PreviewWindow.

-Passed AnimInfo into Automorph so it can be used to set its own
matModel correctly for FO4.

Co-authored-by: ousnius <ousnius@users.noreply.github.com>