Reduced renders from delete-vertices and other commands
- The function wxGLPanel::ShowVertexEdit was being called from a lot
of places it didn't need to be called from.  It only needed to be
called from three places: in MeshesFromProj, to set the bShowPoints
flag for the newly created meshes; in UpdateActiveShapeUI, to reflect
the change in the selection; and in SetVertexEdit, in response to
the user turning the flag on and off.

- The function wxGLPanel::ShowVertexEdit was calling RenderOneFrame.
This was undesirable when being called from RefreshGUIFromProj, as
that function calls it (through MeshesFromProj) before it has finished
setting up the mesh visibility flags, causing a flicker.  So I moved
the call of RenderOneFrame/Render out of ShowVertexEdit and into the
various callers.

- UpdateActiveShapeUI was being called from several places where it
didn't need to be called.  It only needed to be called from one place,
OnShapeSelect, as that is the only function that changes activeItem.

- There were several places where MeshesFromProj or MeshFromProj
were being called directly rather than through RefreshGUIFromProj.
This caused the shape visibility flags to get out of sync with the
GUI, with shapes inappropriately switching to full visibility.
So I changed all such calls to calls of RefreshGUIFromProj.

- MeshFromProj was calling Render.  This was causing two problems:
a lot of extra renders, as MeshFromProj was being called from
a loop in MeshesFromProj; and flickering, as it was called from
RefreshGUIFromProj before the shape visibility flags were restored
(so shapes were briefly rendered visible before being correctly
rendered invisible or wireframe).  So I deleted the call of Render
in MeshFromProj.  RefreshGUIFromProj calls Render itself.

Note that RefreshGUIFromProj is called from a large number of different
places.  It may be overkill in many of those places; I did not check.