Fixed bug in undo-delete-verts with partitions.
The code for undoing deletion of vertices assumed that the
partitions wouldn't have changed.  But empty partitions were being
deleted automatically in SetShapePartitions.  There is no reason to
delete empty partitions here (SetShapeSegments doesn't do so), so I
removed the code for deleting empty partitions from this function.
The code for automatically removing empty partitions is still in
DeleteVertsForShape, for now.