Implemented split-edge command.
Most of the new code is in OutfitProject::PrepareSplitEdge.  It's
pretty complicated because it tries to calculate good values for
positions and diffs that take surface curvature into account somewhat.

I also found that the function wxGLPanel::ShowShape had an unnecessary
call of RenderOneFrame that was causing flickering, so I deleted it.