Temporary fix for split-edge diff corruption issue (#285)
Added code to CollectVertexData and ApplyShapeMeshUndo so that if
the function TargetDataName doesn't return anything, they'll fall
back on concatenating the target name with the slider name.  That
should make it work correctly in most cases.

The main case where it won't work is when some diff data exists for
a slider that is not in activeSet.  For such sets, functions that
add vertices won't work right: split-edge; undo of delete vertices;
undo of separate vertices; and undo of collapse-vertex.  The diff data
won't be corrupted (it'll still be renumbered), but no data will be
added for the vertices.