Sub meshes rendering (#289)
Currently used for rendering partitions and segments separately from one another. Triangles are reordered into a "renderTris" array.

* Implement render logic for sub meshes

Each sub mesh is represented by a start index, a size and optionally a color. This requires the "tris" array to be in a matching order. If there are indices left over that are not part of a sub mesh, they will be rendered as the remainder.

* Made sub-mesh rendering work for partitions and segments.

- Added renderTris to mesh, which is a copy of tris re-ordered for sub-mesh
rendering.

- Added SetSubMeshesForPartitions and SetNoSubMeshes functions to
OutfitStudioFrame.  These set renderTris and subMeshes in mesh.

- Added code to ShowSegment and ShowPartition to set sub-meshes and
sub-mesh colors

* Use color ramp for determining partition/segment colors

Co-authored-by: sts1skj