Fixed bug in SortingMatcher that would sometimes prevent a match.
The problem was this: with the less-than function for the sort, it
was possible to have points A, B, and C such that A<B, B<C, and C<A.
An example:

A: 0, 0, 0
B: -.00007, 1, 0
C: .00007, -1, 0

Because the x values of A and B are within EPSILON of each other, they
are compared using their y values.  The same is true of A and C.  But
B and C have x values more than EPSILON part, so they're compared using
their x values.

So, if two vertices have the coordinates of A, they can be separated in
the sort by B and C.

The older kd_matcher has a similar but less easily triggered problem.
If two vertices have exactly the same coordinates, kd_matcher will always
report a match.  But it can fail if they're not quite the same.  An
example:
A: 0, 1, 0
B: -.00001, 0, 0
C: .00001, 0, 0
B ought to match C, but it won't if A is added first and A's depth
in the tree is such that it's an x-coordinate test node.

I've rewritten SortingMatcher so it should be much more robust.
It doesn't use EPSILON in the sorting step and only sorts on x value,
which requires a lot more work in the matching step.  So this new
SortingMatcher could be a lot slower than the old one for point sets
with lots of close x values.  But it should completely eliminate this
less-than problem.

If it turns out that this new SortingMatcher is too slow, then I
suggest the following:  sort on x.  Identify sequences of points such
that each consecutive pair has x values within EPSILON, and treat
each such sequence as having equal x values (even though the whole
sequence may vary by more than EPSILON).  Then sort each sequence on y,
and do the same for y to produce subsequences with x and y matching.
Then sort the subsequences by z and pick out subsubseqences.